PROD-350 refactor: Fix duplicated components folder
