+++ src/components/BlogCards/BlogCards.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/mixins/fonts.scss";
+@import "../../styles/mixins/fonts.scss";
 
 .content {
   width: 100%;
