+++ src/components/BlogComments/BlogComments.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/mixins/colors.scss";
+@import "../../styles/mixins/colors.scss";
 
 .wrapper {
   form {
