+++ src/components/BlogNewPost/BlogNewPost.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/mixins/fonts.scss";
+@import "../../styles/mixins/fonts.scss";
 
 .main {
   display: grid;
